Clean up Tailwind
Broadly, this commit cleans up a two-year old Tailwind implementation inherited from the corporate app to comport to best practices expressed in Tailwind's documentation and better align with design patterns elsewhere in this app. Specifically, that means...
- Converting Tailwind constants from TS files to JS files
- Consolidating individual named exports to a single default export in tailwind.config.js
- Removing a commented out SCSS import from Tailwind.css
- Removing the utility.css import from main.js because typography.css already imports it
- Installing the dev dependency postcss-import
- Reworking various files such that tailwind.css becomes the single entry point for Tailwind styles
