+++ package.json
@@ -18,6 +18,7 @@
     "@sveltejs/vite-plugin-svelte": "^6.2.4",
     "autoprefixer": "^10.4.19",
     "postcss": "^8.4.39",
+    "postcss-import": "^16.1.1",
     "postcss-prefix-selector": "^2.1.0",
     "svelte": "^5.2.9",
     "svelte-check": "^4.2.2",
