+++ src/styles/tailwind-constants/font-weight.js
@@ -1,4 +1,3 @@
-import type { KeyValuePair, ResolvableTo } from "tailwindcss/types/config";
 import { fontWeight as defaultFontWeight } from "tailwindcss/defaultTheme";
 
 export const fontWeight = {
@@ -10,4 +9,4 @@ export const fontWeight = {
   bold: 700,
   extrabold: 800,
   black: 900,
-} as const satisfies ResolvableTo<KeyValuePair<string, string | number>>;
+};
