+++ src/styles/tailwind-constants/letter-spacing.js
@@ -1,0 +1,10 @@
+import { letterSpacing as defaultLetterSpacing } from "tailwindcss/defaultTheme";
+
+export const letterSpacing = {
+    ...defaultLetterSpacing,
+    tighter: "-0.05em",
+    tight: "-0.01em",
+    normal: "0.00em",
+    wide: "0.01em",
+    wider: "0.03em",
+};
